Update STARTS installation script, add roadmap.md and add STARTS as dependency in pom.xml.

+++ scripts/install-starts.sh
@@ -1,6 +1,10 @@
 git clone https://github.com/TestingResearchIllinois/starts
-cd starts
-git checkout impacted-both-ways
-export JAVA_HOME=$(/usr/libexec/java_home -v1.8)
-mvn install
-rm -rf starts/
+if [ $? -eq 0 ]; then
+  (
+    cd starts
+    git checkout impacted-both-ways
+    export JAVA_HOME=$(/usr/libexec/java_home -v1.8)
+    mvn install
+  )
+  rm -rf starts/
+fi
